Fix `hslToHsv` with 0 saturation and lightness
When given a hsl value with both 0 as the saturation and lightness i.e. black `hsl(0, 0%, 0%)` that would result in a division by 0 in the conversion resulting in `NaN` for the saturation for the hsv result.

This would most easily manifest itself if a swatch were set with the `hsl` value for black, which would generate an invalid color e.g. `#00NANNAN` for the hex representation.