Remove focus on swatch when not active
As the swatch elements are buttons they get the selected border when either focused or having the `pcr-active` class. However changing colors in the picker (via palette etc.), outside of clicking elements that get focus, don’t release the focus on the swatches so they would stay selected.

This change resolves that by only using the `pcr-active` class to include the selected border.