Fix hidden pickr instances issues
If the container the pickr is output in is not visible when it is created the position of the movable handles is not shown correctly on opening, this then has the knock-on effect of not setting the color correctly on interaction unless you click on the palette or a swatch first.

This change addresses that by ensuring the color is set on the first run of `show` after the picker is made visible.

Fixes #226